Set PR to auto-merge
This particular commit is on an orphaned branch where I was trying to
set up the repo so that any push to `dev` would result in a PR, which
would run the Lint-and-Test workflow and then auto-merge if the workflow
succeeded. Unfortunately, this was not successful, because GitHub does
not allow the creation of a PR in an Auto-Merged state, and I can't find
how to generate a PAT which has permissions to make that mutation.

(Further complication would have arisen in the case that, if an initial
commit had failed but then further pushes had fixed it, the long-lived
`dev` branch would appear to have diverged from `main`, because `main`
would have had a squashed-commit on it)

Probably easier - and I will probably do this on the `main`/`dev`
branches - to just have a GitHub action which runs the lint-and-test,
and then (if they have passed) just directly reset `main` to match.